Extract git_scan.py from legacy.py
